+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Parse SQL queries".
 - Tests for "By train or by car ?".
 - Tests for "Low resolution: what's the shape?".
+- Tests for "Simple AI duels".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".